+++ db/column_family.h
@@ -204,6 +204,7 @@ struct SuperVersion {
   // synchronization (ie db mutex held or on write thread).
   ColumnFamilyData* cfd;
   MemTable* mem;
+  MemTable* hot_mem;
   MemTableListVersion* imm;
   Version* current;
   MutableCFOptions mutable_cf_options;
@@ -224,7 +225,7 @@ struct SuperVersion {
   // that needs to be deleted in to_delete vector. Unrefing those
   // objects needs to be done in the mutex
   void Cleanup();
-  void Init(ColumnFamilyData* new_cfd, MemTable* new_mem,
+  void Init(ColumnFamilyData* new_cfd, MemTable* new_mem, MemTable* new_hot_mem,
             MemTableListVersion* new_imm, Version* new_current);
 
   // The value of dummy is not actually used. kSVInUse takes its address as a
@@ -348,13 +349,12 @@ class ColumnFamilyData {
   InternalStats* internal_stats() { return internal_stats_.get(); }
 
   MemTableList* imm() { return &imm_; }
-  MemTableList* hot_imm() { return &hot_imm_; }
   MemTable* mem() { return mem_; }
   MemTable* hot_mem() {return hot_mem_; }
 
   bool IsEmpty() {
     return mem()->GetFirstSequenceNumber() == 0 && imm()->NumNotFlushed() == 0 
-           && hot_mem()->GetFirstSequenceNumber() == 0 && hot_imm()->NumNotFlushed() == 0;
+           && hot_mem()->GetFirstSequenceNumber() == 0 ;
   }
 
   Version* current() { return current_; }
@@ -586,7 +586,6 @@ class ColumnFamilyData {
   MemTable* mem_;
   MemTable* hot_mem_;
   MemTableList imm_;
-  MemTableList hot_imm_;
   SuperVersion* super_version_;
 
   // An ordinal representing the current SuperVersion. Updated by